+++ CHANGELOG.md
@@ -9,6 +9,7 @@
  * Builtin functions to check for nan, infinity and subnormality in floats (#101)
  * Builtin random function (#102)
  * Implement `TryFrom<Value>` for all types a value can hold (#105)
+ * Split VariableIdentifier node into read and write variants (#106)
 
 ### Removed
 
@@ -24,6 +25,7 @@ My warmhearted thanks goes to:
 
  * [Ophir LOJKINE](https://github.com/lovasoa)
  * [Joe Grund](https://github.com/jgrund)
+ * [Luka Maljic](https://github.com/malj)
 
 ## [7.2.0](https://github.com/ISibboI/evalexpr/compare/7.1.1...7.2.0) - 2022-03-16